# docs/04-api/Enterprise_AI_API_Gateway_OpenAPI_3.1_V1.2_Final.yaml
openapi: 3.1.0
info:
  title: Enterprise AI & API Gateway
  version: 1.2.0
  summary: Final executable HTTP contract for Enterprise AI & API Gateway V1
  description: Derived from PRD V1.4, HLD V1.1 corrections, ERD V1.2 Final, Physical DB/Alembic V1.2 Final, API Specification
    V1.2.2 Final, frozen M0-M9 execution contracts, and the M2.6 Normal API Registry Contract Reconciliation addendum.
    This artifact consolidates final V1 semantics for Normal API, OpenAI-compatible LLM, Admin browser sessions,
    traffic-admission health, idempotency, certification, credential invalidation, SSRF/TLS and SSE.
servers:
- url: https://gateway.example.com
  description: Production example
security: []
...
paths:
...
  /api/v1/admin/api-keys/{api_key_id}/permissions:
    parameters:
    - $ref: '#/components/parameters/ApiKeyId'
    get:
      tags:
      - API Keys
      operationId: listApiKeyPermissions
      security:
      - AdminToken: []
      responses:
        '200':
          description: Permission list
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ApiKeyPermission'
    put:
      tags:
      - API Keys
      operationId: replaceApiKeyPermissions
      security:
      - AdminToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - permissions
              additionalProperties: false
              properties:
                permissions:
                  type: array
                  items:
                    $ref: '#/components/schemas/ApiKeyPermissionWrite'
      responses:
        '200':
          description: Permission set replaced atomically
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
      parameters:
      - $ref: '#/components/parameters/CsrfToken'
      description: When authenticated by AdminSessionCookie, X-CSRF-Token is mandatory. Supplying both AdminToken and AdminSessionCookie
        is rejected as ambiguous credentials.
...
components:
  securitySchemes:
    GatewayApiKey:
      type: http
      scheme: bearer
      bearerFormat: gw_<secret>
      description: Data Plane credential only.
    AdminToken:
      type: http
      scheme: bearer
      bearerFormat: adm_<secret>
      description: Per-admin Control Plane credential only.
    AdminSessionCookie:
      type: apiKey
      in: cookie
      name: gw_admin_session
      description: Opaque Redis-backed Admin browser session cookie. Fail-closed when session store is unavailable.
  parameters:
...
    ApiKeyId:
      name: api_key_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: false
      description: Required only when the operation is authenticated with AdminSessionCookie and the HTTP method is state-changing.
        Not required for direct Bearer adm_ authentication.
      schema:
        type: string
        minLength: 16
        maxLength: 512
...
  responses:
...
    NotFound:
      description: Not found; cross-tenant references also return 404
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InvalidRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
      - error
      - request_id
      additionalProperties: false
      properties:
        error:
          type: object
          required:
          - message
          - type
          - code
          additionalProperties: false
          properties:
            message:
              type: string
            type:
              type: string
              enum:
              - authentication_error
              - authorization_error
              - rate_limit_exceeded
              - invalid_request
              - resource_not_found
              - model_not_found
              - model_not_allowed
              - provider_timeout
              - provider_rate_limit
              - provider_unavailable
              - provider_authentication_error
              - provider_error
              - all_targets_unavailable
              - gateway_timeout
              - gateway_internal_error
              - resource_conflict
              - route_conflict
              - active_credential_conflict
              - price_window_conflict
              - idempotency_conflict
              - configuration_conflict
              - request_too_large
              - upstream_unavailable
              - upstream_timeout
              - idempotency_in_progress
            param:
              type:
              - string
              - 'null'
            code:
              type: string
            metadata:
              type: object
              additionalProperties: true
              description: Safe aggregate diagnostics only; no provider IDs, base URLs, credentials, or topology.
              properties:
                attempted_targets:
                  type: integer
                  minimum: 0
                ineligible_targets:
                  type: integer
                  minimum: 0
                last_error_code:
                  type:
                  - string
                  - 'null'
        request_id:
          type: string
          format: uuid
          description: Canonical Gateway-generated UUID v7.
...
    ApiKeyPermissionWrite:
      type: object
      required:
      - resource_type
      - resource_id
      - action
      additionalProperties: false
      properties:
        resource_type:
          type: string
          enum:
          - SERVICE
          - ROUTE
          - LLM_ALIAS
          - LLM_MODEL
        resource_id:
          type: string
          format: uuid
        action:
          type: string
          const: INVOKE
    ApiKeyPermission:
      allOf:
      - $ref: '#/components/schemas/ApiKeyPermissionWrite'
      - type: object
        properties:
          id:
            type: string
            format: uuid
...